fastapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-keycloak==4.7.3
pydantic-settings==2.6.1
email-validator==2.2.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
